trihard/requirements.txt
numpy>=2.1.0
pandas>=2.2.2
matplotlib>=3.9.2
progressbar2>=4.4.2